# docs/swagger.yaml
basePath: /api
definitions:
  courseerror.CourseError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  entity.Passwords:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
...
  entity.SuccessResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  entity.UserBilling:
    properties:
      id:
        type: integer
      invoice:
        type: integer
      order:
        type: string
      paid:
        type: number
      paidStatus:
        type: boolean
      paymentMethod:
        type: string
      timestamp:
        type: string
    type: object
  entity.UserCourses:
    properties:
      billingInfo:
        items:
          $ref: '#/definitions/entity.UserBilling'
        type: array
      id:
        type: integer
      name:
        type: string
      previewUrl:
        type: string
    type: object
  entity.UserData:
    properties:
      courses:
        items:
          $ref: '#/definitions/entity.UserCourses'
        type: array
      email:
        type: string
      firstName:
        type: string
      isEmailVerified:
        type: boolean
      phoneNumber:
        type: integer
      photo:
        type: string
      surname:
        type: string
    type: object
...
  entity.UserInfo:
    properties:
      firstName:
        type: string
      phoneNumber:
        type: string
      surname:
        type: string
    type: object
...
host: localhost:70
info:
  contact: {}
  title: Приложение course
  version: "1.0"
paths:
...
  /v1/profile/confirmEmailChange:
    post:
      parameters:
      - description: Код подтверждения
        in: query
        name: confirmCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Подтвердить изменении почты пользователя
      tags:
      - Методы для администрирования профиля
  /v1/profile/disable:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Ошибка получения userId
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Заморозить профиль
      tags:
      - Методы для администрирования профиля
  /v1/profile/editEmail:
    patch:
      parameters:
      - description: Почта
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Изменить почту пользователя
      tags:
      - Методы для администрирования профиля
  /v1/profile/editPassword:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Пароли
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/entity.Passwords'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация или не удалось декодировать сообщение
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Изменить пароль пользователя
      tags:
      - Методы для администрирования профиля
  /v1/profile/editProfile:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Данные профиля
        in: body
        name: UserInfo
        required: true
        schema:
          $ref: '#/definitions/entity.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация или не удалось декодировать сообщение
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Изменить профиль пользователя
      tags:
      - Методы для администрирования профиля
...
  /v1/profile/getUser:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserData'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Получить данные профиля
      tags:
      - Методы для администрирования профиля
...
  /v1/profile/setPhoto:
    patch:
      consumes:
      - multipart/form-data
      parameters:
      - description: Фото
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация или не получилось обработать фото
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Изменить фото профиля
      tags:
      - Методы для администрирования профиля
  /v1/profile/watchLesson:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Ошибка получения userId
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Урок не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Пометить урок как пройденный
      tags:
      - Методы для администрирования профиля
securityDefinitions:
  AdminApiKeyAuth:
    in: cookie
    name: admin_auth
    type: apiKey
  ApiKeyAuth:
    in: cookie
    name: auth
    type: apiKey
swagger: "2.0"